Refactor into separate module
