+++ looper/command_models/__init__.py
@@ -1,0 +1,6 @@
+"""
+This module holds `pydantic` models that describe commands and their arguments.
+
+These can be used either with the `pydantic-argparse` library to build a CLI or
+by an HTTP API.
+"""
